feat: add authorizing verification at project,bucket,note

+++ router/dashboard/note.py
@@ -4,6 +4,7 @@ from pydantic import UUID4
 
 from database import schemas
 from func.dashboard.crud.note import *
+from func.auth.auth import *
 
 
 router = APIRouter(
@@ -17,6 +18,11 @@ router = APIRouter(
 
 @router.get("/{note_id}", tags=["note"])
 async def get_note(req: Request, note_id: str):
+    user: UUID4 = verify_user(req)
+    # if not user:
+    #     raise HTTPException(status_code=401, detail="Unauthorized")
+    if not verify_note(user, note_id):
+        raise HTTPException(status_code=401, detail="Unauthorized")
     res = read_note(note_id)
     return JSONResponse(content={
         "status": "succeed",
@@ -28,6 +34,11 @@ async def get_note(req: Request, note_id: str):
 
 @router.get("/list/{bucket_id}", tags=["note"])
 async def get_note_list(req: Request, bucket_id: str):
+    user: UUID4 = verify_user(req)
+    # if not user:
+    #     raise HTTPException(status_code=401, detail="Unauthorized")
+    if not verify_bucket(user, bucket_id):
+        raise HTTPException(status_code=401, detail="Unauthorized")
     res = read_note_list(bucket_id)
     return JSONResponse(content={
         "status": "succeed",
@@ -39,6 +50,14 @@ async def get_note_list(req: Request, bucket_id: str):
 
 @router.post("/", tags=["note"])
 async def add_note(req: Request, note: schemas.NoteCreate):
+    user: UUID4 = verify_user(req)
+    # if not user:
+    #     raise HTTPException(status_code=401, detail="Unauthorized")
+    if not verify_bucket(user, note["bucket_id"]):
+        raise HTTPException(status_code=401, detail="Unauthorized")
+    
+    # need verify timestamp logic
+
     res = create_note(note)
     return JSONResponse(content={
         "status": "succeed",
@@ -50,6 +69,14 @@ async def add_note(req: Request, note: schemas.NoteCreate):
 
 @router.put("/{note_id}", tags=["note"])
 async def change_note(req: Request, note: schemas.NoteUpdate):
+    user: UUID4 = verify_user(req)
+    # if not user:
+    #     raise HTTPException(status_code=401, detail="Unauthorized")
+    if not user == note["user_id"]:
+        raise HTTPException(status_code=401, detail="Unauthorized")
+    
+    # need verify timestamp logic
+
     res = update_note(note)
     return JSONResponse(content={
         "status": "succeed",
@@ -61,6 +88,14 @@ async def change_note(req: Request, note: schemas.NoteUpdate):
 
 @router.delete("/{note_id}", tags=["note"])
 async def drop_note(req: Request, note_id: str):
+    user: UUID4 = verify_user(req)
+    # if not user:
+    #     raise HTTPException(status_code=401, detail="Unauthorized")
+    data, count = supabase.table("note").select("user_id").eq("id", note_id).execute()
+    if not data[1]:
+        raise HTTPException(status_code=400, detail="No data")
+    if not user == data[1][0]["user_id"]:
+        raise HTTPException(status_code=401, detail="Unauthorized")
     res = delete_note(note_id)
     return JSONResponse(content={
         "status": "succeed",
